requirements.txt
streamlit==1.46.1
google-genai==1.24.0
